Dw 5791 (#590)
* fixing fetching releases


* fixing var names


* fix for jar


* fix typo


* execute update tfvars at the end


* fix var names


* fix name again


* emr realuncher name fix


* typo


* reset ci things


* reset things


* fix exporter naming


* reset tf vars


* put some things back to normal

+++ get_lambda_release.sh
@@ -1,7 +1,44 @@
 #!/usr/bin/env bash
-VERSION=$(curl --silent "https://api.github.com/repos/dwp/${REPO}/releases/latest" | grep '"tag_name":' | sed -E 's/.*"([^"]+)".*/\1/')
-echo "my version is $VERSION"
-echo "$(echo ${REPO} | sed 's/-/_/g')_zip = { base_path = \"../${REPO}-release\", version = \"${VERSION}\" }" >> terraform.tfvars
-mkdir ../${REPO}-release
-ASSET=$(curl --silent "https://api.github.com/repos/dwp/${REPO}/releases/latest" | grep "browser_download_url.*zip" | cut -d : -f 2,3 | tr -d \")
-(cd ../${REPO}-release/ && curl -L -O ${ASSET})
+
+get_release_information(){
+    VERSION=$(curl --silent "https://api.github.com/repos/dwp/${REPO}/releases/latest" | grep '"tag_name":' | sed -E 's/.*"([^"]+)".*/\1/')
+    RESPONSE=`echo $(curl --silent "https://api.github.com/repos/dwp/${REPO}/releases/latest")`
+}
+
+update_tfvars() {
+    tfvars_line=`echo "$(echo ${REPO} | sed 's/-/_/g')_$extension = { base_path = \"../${REPO}-release\", version = \"${VERSION}\" }"`
+    if ! cat terraform.tfvars | grep -q "$tfvars_line"; then
+        echo $tfvars_line >> terraform.tfvars
+    fi
+}
+
+get_release() {
+    get_release_information
+    for k in $(jq '.assets | keys | .[]' <<< "$RESPONSE"); do
+        value=$(jq -r ".assets[$k]" <<< $RESPONSE);
+        url=$(jq -r ".browser_download_url" <<< $value);
+        RELEASE=${REPO}-${VERSION}
+        if  echo "$url" | grep -q "$RELEASE"; then
+            export ASSET=$url
+            filename=$(basename -- "$ASSET")
+            extension="${filename##*.}"
+            fetch_asset
+        fi
+    done
+}
+
+fetch_asset(){
+    EXISTING_VERSION=`ls ../${REPO}-release/ | grep ${VERSION}`
+
+    if [[ -f ../${REPO}-release/$EXISTING_VERSION ]]; then
+        echo "${REPO}-release/$EXISTING_VERSION already exists: Skipping download"
+    else
+        mkdir ../${REPO}-release
+        (cd ../${REPO}-release/ && curl -L -O ${ASSET})
+    fi
+
+    update_tfvars
+}
+
+get_release
+
